# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for LIWACH system",
        "description": "API Documentation for LIWACH system",
        "contact": {
            "email": "edenabdisa90@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://liwachapi.herokuapp.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/reporttype": {
            "get": {
                "tags": [
                    "ReportType"
                ],
                "summary": "Get list of reporttype",
                "description": "Returns list of reporttype",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "ReportType"
                ],
                "summary": "Store new ReportType",
                "description": "Returns ReportType data",
                "operationId": "storeReportType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/reporttype/{id}": {
            "put": {
                "tags": [
                    "ReportType"
                ],
                "summary": "Update existing ReportType",
                "description": "Returns updated ReportType data",
                "operationId": "updatePReportType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReportType id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReportType"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ReportType"
                ],
                "summary": "Delete existing ReportType",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteReportType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ReportType id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReportType": {
                "title": "ReportType",
                "description": "Report type model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "big integer",
                        "format": "int64",
                        "example": 1
                    },
                    "report_detail": {
                        "title": "Name",
                        "description": "Small text that will be displayed to user so that they can choose from the list",
                        "type": "string",
                        "example": "Spam, Fake Item, Violence, Child Abuse, Other, "
                    },
                    "type_for": {
                        "title": "Type for",
                        "description": "The item that this report can be associated with",
                        "type": "string",
                        "example": "item or service or user"
                    },
                    "timestamp": {
                        "title": "Timestamp",
                        "description": "This is the timestamp that we got when this data is added.",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                }
            },
            "ReportTypeRequest": {
                "title": "ReportType",
                "description": "Report type model",
                "properties": {
                    "report_detail": {
                        "title": "Name",
                        "description": "Small text that will be displayed to user so that they can choose from the list",
                        "type": "string",
                        "example": "Spam, Fake Item, Violence, Child Abuse, Other, "
                    },
                    "type_for": {
                        "title": "Type for",
                        "description": "The item that this report can be associated with",
                        "type": "string",
                        "example": "item or service or user"
                    }
                }
            },
...
        }
    }
}